+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+- Reply reviews
+
 ## [2.2.1] - 2020-10-27
 ### Fixed
 - Missing parenthesis in translation message.